+++ lib.php
@@ -479,8 +479,11 @@ function board_cm_info_dynamic(cm_info $cm) {
     // Look up the board based on the course module.
     $board = board::get_board($cm->instance);
 
+    // Check if embedding feature is allowed.
+    $embedallowed = get_config('mod_board', 'embed_allowed');
+
     // If we are embedding the board, turn off the view link.
-    if ($board->embed) {
+    if ($embedallowed && $board->embed) {
         $cm->set_no_view_link();
     }
 
@@ -496,10 +499,16 @@ function board_cm_info_view(cm_info $cm) {
     // Look up the board based on the course module.
     $board = board::get_board($cm->instance);
 
-    if ($board->embed) {
+    // Check if embedding feature is allowed.
+    $embedallowed = get_config('mod_board', 'embed_allowed');
+
+    if ($embedallowed && $board->embed) {
         $width = get_config('mod_board', 'embed_width');
         $height = get_config('mod_board', 'embed_height');
-        $output = html_writer::tag('h3', $board->name);
+        $output = html_writer::start_tag('div', ['class' => 'mod_board_embed_container']);
+        if (empty($board->hidename)) {
+            $output .= html_writer::tag('h3', $board->name);
+        }
         $output .= html_writer::start_tag('iframe', [
             'src' => new moodle_url('/mod/board/view.php', ['id' => $cm->id, 'embed' => 1]),
             'width' => $width,
@@ -510,6 +519,7 @@ function board_cm_info_view(cm_info $cm) {
         $output .= html_writer::end_tag('iframe');
         $output .= html_writer::link(new moodle_url('/mod/board/view.php', ['id' => $cm->id]),
             get_string('viewboard', 'board'));
+        $output .= html_writer::end_tag('div');
         $cm->set_content($output, true);
     }
 }