+++ settings.php
@@ -67,6 +67,13 @@ if ($ADMIN->fulltree) {
         '1'
     ));
 
+    $settings->add(new admin_setting_configcheckbox(
+        'mod_board/embed_allowed',
+        get_string('embed_allowed', 'mod_board'),
+        get_string('embed_allowed_desc', 'mod_board'),
+        '1'
+    ));
+
     $settings->add(new admin_setting_configtext('mod_board/embed_width', get_string('embed_width', 'mod_board'),
                        get_string('embed_width_desc', 'mod_board'), '99%', PARAM_TEXT));
 
@@ -117,5 +124,48 @@ if ($ADMIN->fulltree) {
             board::SINGLEUSER_PRIVATE => get_string('singleusermodeprivate', 'mod_board'),
             board::SINGLEUSER_PUBLIC => get_string('singleusermodepublic', 'mod_board'),
         ]
+    ));
+
+    // Heading.
+    $setting = new admin_setting_heading('mod_board/settings_heading_logging',
+            get_string('settings_heading_logging', 'mod_board'),
+            get_string('settings_heading_logging_info', 'mod_board'));
+    $settings->add($setting);
+
+    $settings->add(new admin_setting_configcheckbox(
+            'mod_board/addcolumnnametolog',
+            get_string('settings:addcolumnnametolog', 'mod_board'),
+            '',
+            '1'
+    ));
+    $settings->add(new admin_setting_configcheckbox(
+            'mod_board/addnotetolog',
+            get_string('settings:addnotetolog', 'mod_board'),
+            '',
+            '1'
+    ));
+    $settings->add(new admin_setting_configcheckbox(
+            'mod_board/addcommenttolog',
+            get_string('settings:addcommenttolog', 'mod_board'),
+            '',
+            '1'
+    ));
+    $settings->add(new admin_setting_configcheckbox(
+            'mod_board/addheadingtolog',
+            get_string('settings:addheadingtolog', 'mod_board'),
+            '',
+            '1'
+    ));
+    $settings->add(new admin_setting_configcheckbox(
+            'mod_board/addattachmenttolog',
+            get_string('settings:addattachmenttolog', 'mod_board'),
+            '',
+            '1'
+    ));
+    $settings->add(new admin_setting_configcheckbox(
+            'mod_board/addratingtolog',
+            get_string('settings:addratingtolog', 'mod_board'),
+            '',
+            '1'
     ));
 }